+++ unn/algo/graph_lab/GraphLab/GraphLab/GraphSPT.h
@@ -1,4 +1,4 @@
 #include "Graph.h"
 
-std::vector<double> Deijkstra_SPT(const Graph& g, int s);
-std::vector<double> FordBellman_SPT(const Graph& g, int s);
+std::vector<double> Deijkstra_SPT(const Graph& g, size_t s);
+std::vector<double> FordBellman_SPT(const Graph& g, size_t s);
